+++ example/lib/main.dart
@@ -28,12 +28,14 @@ class _TabbedViewExamplePageState extends State<TabbedViewExamplePage> {
 
     tabs.add(TabData(
         text: 'Tab 1',
+        leading: (context, status) => Icon(Icons.star, size: 16),
         content: Padding(child: Text('Hello'), padding: EdgeInsets.all(8))));
     tabs.add(TabData(
         text: 'Tab 2',
         content:
             Padding(child: Text('Hello again'), padding: EdgeInsets.all(8))));
-    tabs.add(TabData(closable: false,
+    tabs.add(TabData(
+        closable: false,
         text: 'TextField',
         content: Padding(
             child: TextField(
@@ -47,10 +49,9 @@ class _TabbedViewExamplePageState extends State<TabbedViewExamplePage> {
 
   @override
   Widget build(BuildContext context) {
-
     TabbedView tabbedView = TabbedView(controller: _controller);
-    Widget w = TabbedViewTheme(child: tabbedView, data: TabbedViewThemeData.mobile());
-    return Scaffold(
-        body: Container(child: w, padding: EdgeInsets.all(32)));
+    Widget w =
+        TabbedViewTheme(child: tabbedView, data: TabbedViewThemeData.mobile());
+    return Scaffold(body: Container(child: w, padding: EdgeInsets.all(32)));
   }
 }